backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "google-api-python-client>=2.167.0",
    "google-auth>=2.39.0",
    "google-auth-oauthlib>=1.2.2",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "uvicorn[standard]>=0.34.2",
]
